Move date to it's own utility file
