+++ scripts/lib/CIME/case_st_archive.py
@@ -6,7 +6,8 @@ import shutil, glob, re, os
 
 from CIME.XML.standard_module_setup import *
 from CIME.case_submit               import submit
-from CIME.utils                     import run_and_log_case_status, ls_sorted_by_mtime, symlink_force, date
+from CIME.utils                     import run_and_log_case_status, ls_sorted_by_mtime, symlink_force
+from CIME.date                      import date
 from os.path                        import isdir, join
 
 logger = logging.getLogger(__name__)
